// src/types/enums.ts
export enum CardNewsType {
  top = '', //564px
  block = '430px',
};

export enum ButtonType {
  signin = 'custombutton__signin',
  signup = 'custombutton__signup',
  google = 'custombutton__google',
  facebook = 'custombutton__facebook',
  apple = 'custombutton__apple',
  subscribe_yellow = 'custombutton__subscribe-yellow',
  subscribe_blue = 'custombutton__subscribe-blue',
  subscribe_purple = 'custombutton__subscribe-purple',
};

export enum PlanType {
  lite = 'yellow',
  basic = 'blue',
  premium = 'purple',
};

export enum ModalType {
  signin,
  signup,
  forgotpass,
  resetpass,
  checkmail,
};